chore: extract insufficient balance check to hook
