+++ app/components/UI/Bridge/components/TokenInputArea/index.tsx
@@ -37,6 +37,7 @@ import { getDisplayCurrencyValue } from '../../utils/exchange-rates';
 import { useBridgeExchangeRates } from '../../hooks/useBridgeExchangeRates';
 import { useLatestBalance } from '../../hooks/useLatestBalance';
 import { parseUnits } from 'ethers/lib/utils';
+import useIsInsufficientBalance from '../../hooks/useInsufficientBalance';
 
 const createStyles = () =>
   StyleSheet.create({
@@ -147,24 +148,8 @@ export const TokenInputArea = forwardRef<
     const networkConfigurationsByChainId = useSelector(
       selectNetworkConfigurations,
     );
-    const { quoteRequest } = useSelector(selectBridgeControllerState);
 
-    const latestBalance = useLatestBalance({
-      address: token?.address,
-      decimals: token?.decimals,
-      chainId: token?.chainId,
-      balance: token?.balance,
-    });
-    const isValidAmount =
-    amount !== undefined && amount !== '.' && token?.decimals;
-
-    // quoteRequest.insufficientBal is undefined for Solana quotes, so we need to manually check if the source amount is greater than the balance
-    const isInsufficientBalance =
-    quoteRequest?.insufficientBal ||
-    (isValidAmount &&
-      parseUnits(amount, token.decimals).gt(
-        latestBalance?.atomicBalance ?? BigNumber.from(0),
-      ));
+    const isInsufficientBalance = useIsInsufficientBalance({ amount, token });
 
     let nonEvmMultichainAssetRates = {};
     ///: BEGIN:ONLY_INCLUDE_IF(keyring-snaps)
